+++ test/setup_helpers.go
@@ -1,4 +1,4 @@
-package testutils
+package test
 
 import "net"
