+++ test/utils.go
@@ -1,4 +1,4 @@
-package testutils
+package test
 
 import (
 	"fmt"
